httpclient: Be more strict about the GMainContext doing the clean up
The SoupSession is meant to be used with the thread-default
GMainContext at the time of its creation. Destroying it from anywhere
else will lead to crashes or deadlocks. Given that we are only using
asynchronous operations, chances are very high that the thread-default
context is the same in the SoupSessionCallback.

However, let's not leave it to chance. Unlike GSimpleAsyncResult,
GTask offers us a way to be absolutely sure. Let's take advantage of
it.

https://bugzilla.gnome.org/show_bug.cgi?id=743044
https://bugzilla.gnome.org/show_bug.cgi?id=764157